+++ app/Livewire/Sections/Delete.php
@@ -4,6 +4,7 @@ namespace App\Livewire\Sections;
 
 use Illuminate\Support\Facades\Storage;
 use App\Models\Section;
+use Illuminate\Support\Facades\Gate;
 use Livewire\Component;
 
 class DeleteSection extends Component
@@ -12,7 +13,7 @@ class DeleteSection extends Component
 
     public function delete($id, $photo)
     {
-
+        Gate::authorize('section.delete');
         Section::destroy($id);
 
         Storage::disk('public')->delete('uploads/' . $photo);
